requirements.txt
absl-py                      2.0.0
anyio                        4.0.0
argon2-cffi                  23.1.0
argon2-cffi-bindings         21.2.0
arrow                        1.3.0
asttokens                    2.4.0
astunparse                   1.6.3
async-lru                    2.0.4
attrs                        23.1.0
Babel                        2.13.0
backcall                     0.2.0
beautifulsoup4               4.12.2
bleach                       6.1.0
blinker                      1.6.3
cachetools                   5.3.2
certifi                      2023.7.22
cffi                         1.16.0
charset-normalizer           3.3.0
click                        8.1.7
colorama                     0.4.6
comm                         0.1.4
contourpy                    1.1.1
cycler                       0.12.1
debugpy                      1.8.0
decorator                    5.1.1
defusedxml                   0.7.1
exceptiongroup               1.1.3
executing                    2.0.0
fastjsonschema               2.18.1
filelock                     3.12.4
Flask                        3.0.0
flatbuffers                  23.5.26
fonttools                    4.43.1
fqdn                         1.5.1
fsspec                       2023.9.2
gast                         0.5.4
gitdb                        4.0.10
GitPython                    3.1.38
google-auth                  2.23.3
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.59.0
h5py                         3.10.0
idna                         3.4
ipykernel                    6.25.2
ipython                      8.16.1
isoduration                  20.11.0
itsdangerous                 2.1.2
jedi                         0.19.1
Jinja2                       3.1.2
joblib                       1.3.2
json5                        0.9.14
jsonpointer                  2.4
jsonschema                   4.19.1
jsonschema-specifications    2023.7.1
jupyter_client               8.4.0
jupyter_core                 5.4.0
jupyter-events               0.8.0
jupyter-lsp                  2.2.0
jupyter_server               2.8.0
jupyter-server-mathjax       0.2.6
jupyter_server_terminals     0.4.4
jupyterlab                   4.0.7
jupyterlab_git               0.43.0
jupyterlab-pygments          0.2.2
jupyterlab_server            2.25.0
kaggle                       1.5.16
keras                        2.14.0
kiwisolver                   1.4.5
libclang                     16.0.6
Markdown                     3.5
MarkupSafe                   2.1.3
matplotlib                   3.8.0
matplotlib-inline            0.1.6
mistune                      3.0.2
ml-dtypes                    0.2.0
mpmath                       1.3.0
nbclient                     0.8.0
nbconvert                    7.9.2
nbdime                       3.2.1
nbformat                     5.9.2
nest-asyncio                 1.5.8
networkx                     3.1
notebook_shim                0.2.3
numpy                        1.26.1
nvidia-cublas-cu12           12.1.3.1
nvidia-cuda-cupti-cu12       12.1.105
nvidia-cuda-nvrtc-cu12       12.1.105
nvidia-cuda-runtime-cu12     12.1.105
nvidia-cudnn-cu12            8.9.2.26
nvidia-cufft-cu12            11.0.2.54
nvidia-curand-cu12           10.3.2.106
nvidia-cusolver-cu12         11.4.5.107
nvidia-cusparse-cu12         12.1.0.106
nvidia-nccl-cu12             2.18.1
nvidia-nvjitlink-cu12        12.2.140
nvidia-nvtx-cu12             12.1.105
oauthlib                     3.2.2
opt-einsum                   3.3.0
overrides                    7.4.0
packaging                    23.2
pandas                       2.1.1
pandocfilters                1.5.0
parso                        0.8.3
pexpect                      4.8.0
pickleshare                  0.7.5
Pillow                       10.1.0
pip                          23.3
platformdirs                 3.11.0
plotly                       5.17.0
prometheus-client            0.17.1
prompt-toolkit               3.0.39
protobuf                     4.24.4
psutil                       5.9.6
ptyprocess                   0.7.0
pure-eval                    0.2.2
pyasn1                       0.5.0
pyasn1-modules               0.3.0
pycparser                    2.21
Pygments                     2.16.1
pyparsing                    3.1.1
python-dateutil              2.8.2
python-json-logger           2.0.7
python-slugify               8.0.1
pytz                         2023.3.post1
PyYAML                       6.0.1
pyzmq                        25.1.1
referencing                  0.30.2
requests                     2.31.0
requests-oauthlib            1.3.1
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rpds-py                      0.10.6
rsa                          4.9
scikit-learn                 1.3.1
scipy                        1.11.3
seaborn                      0.13.0
Send2Trash                   1.8.2
setuptools                   68.0.0
six                          1.16.0
smmap                        5.0.1
sniffio                      1.3.0
soupsieve                    2.5
stack-data                   0.6.3
sympy                        1.12
tenacity                     8.2.3
tensorboard                  2.14.1
tensorboard-data-server      0.7.2
tensorflow                   2.14.0
tensorflow-estimator         2.14.0
tensorflow-io-gcs-filesystem 0.34.0
termcolor                    2.3.0
terminado                    0.17.1
text-unidecode               1.3
threadpoolctl                3.2.0
tinycss2                     1.2.1
tomli                        2.0.1
torch                        2.1.0
tornado                      6.3.3
tqdm                         4.66.1
traitlets                    5.11.2
triton                       2.1.0
types-python-dateutil        2.8.19.14
typing_extensions            4.8.0
tzdata                       2023.3
uri-template                 1.3.0
urllib3                      2.0.7
wcwidth                      0.2.8
webcolors                    1.13
webencodings                 0.5.1
websocket-client             1.6.4
Werkzeug                     3.0.1
wheel                        0.41.3
wrapt                        1.14.1